% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dod-package.R, R/dod.R
\docType{package}
\name{dod}
\alias{dod}
\title{dod:}
\usage{
dod(surv_ts, dod_model, timepoints, learning_type="unsupervised", maxIter=100, verbose=FALSE)
}
\arguments{
\item{surv_ts}{Either a surveillance time series (sts) object created using the surveillance package or a list of sts objects.}

\item{dod_model}{An object of class \code{\linkS4class{DODmodel}} which specifies the model parameters and structure}

\item{timepoints}{Specifying whether confidence intervals should be computed, default: TRUE.}

\item{learning_type}{Indicates the type of learning, one of c("unsupervised", "semisupervised", "supervised")}

\item{maxIter}{Maximal number of iteration for EM-algorithm.}

\item{verbose}{Logical indicating wehther progress should be printed.}
}
\description{
The \strong{dod} package

Disease outbreak detection - package main function
}
\section{dod functions}{

\code{\link{dod}}:
}

\examples{

TODO
  
}
\seealso{
\code{\linkS4class{DODmodel}}
}
