% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-DODfamily-class.R
\name{DODfamily}
\alias{DODfamily}
\title{Create a family of probability distributions for outbreak detection}
\usage{
DODfamily(name, shared_dispersion=TRUE, shared_pi=TRUE)
}
\arguments{
\item{name}{Name of the probability distribution that should be used. One of c("Poisson", "ZIPoisson", "NegBinom", "ZINegBinom").}

\item{shared_dispersion}{Logical indicating whether dispersion parameter is shared across multiple time series. Only relevant if name is one of c("NegBinom", "ZINegBinom").}

\item{shared_pi}{Logical indicating whether parameter pi is shared across multiple time series. Only relevant if name is one of c("ZIPoisson", "ZINegBinom").}
}
\description{
Create a family of probability distributions for outbreak detection
}
\examples{

TODO
  
}
\seealso{
\code{\linkS4class{DODfamily}}
}
