% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-DODformula-class.R
\docType{class}
\name{DODformula-class}
\alias{DODformula-class}
\title{This class is a generic container for a formula which specifies the model used for outbreak detection}
\description{
This class is a generic container for a formula which specifies the model used for outbreak detection
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the model. Must be one of c("FarringtonNoufaily", "Harmonic", "ExtData").}

\item{\code{shared_params}}{Indicates whether model parameters are shared across multiple time series.}
}}

