% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-DODformula-class.R
\name{DODformula}
\alias{DODformula}
\title{Create a formula of the model for outbreak detection}
\usage{
DODformula(name, S=1, freq=52, noPeriods=10, w=3, timeTrend=TRUE, extdata=NULL, shared_params=TRUE, offset=FALSE)
}
\arguments{
\item{name}{Name of the model. Must be one of c("FarringtonNoufaily", "Harmonic", "ExtData").}

\item{S}{Number of oscillations during one year.}

\item{freq}{Number of time units (can be e.g. weeks or months) in a year.}

\item{noPeriods}{Number of levels in the factor which creates bins in each year to model seasonal patterns.}

\item{w}{The number of weeks before and after the current week to include in the bin which contains the respective week in each year.}

\item{timeTrend}{Indicates whether a time trend should be included in the model.}

\item{extdata}{A data.frame containing variables which are used to model case counts. All variables in the data.frame will be used in the model. The data.frame has to have the same number of rows as time points in the time series.}

\item{shared_params}{Indicates whether model parameters are shared across multiple time series.}

\item{offset}{An offset to included in the predictor during fitting.}
}
\description{
Create a formula of the model for outbreak detection
}
\examples{

TODO
  
}
\seealso{
\code{\linkS4class{DODformula}}
}
