% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DODmodel-class.R
\docType{class}
\name{DODmodel-class}
\alias{DODmodel-class}
\title{This class is a generic container for the model used for outbreak detection}
\description{
This class is a generic container for the model used for outbreak detection
}
\section{Slots}{

\describe{
\item{\code{emission}}{Emission of hidden Markov model. Includes a DODfamily and a DODformula object.}

\item{\code{transitions}}{Transition proabilities of hidden Markov model.}

\item{\code{transitions_prior}}{Indicates whether a prior for transition probabilities should be used.}

\item{\code{prior_weights}}{Dirichlet prior weights for transition probabilities. This cannot be changed by the user.}

\item{\code{loglik_transitions}}{Prior log likelihood of transition probabilities. Only relevant for model fitting.}

\item{\code{initial_prob}}{Initial state probabilities of the hidden Markov model.}

\item{\code{setBckgState}}{Indicates whether a background state should be computed for model fitting. Background states are set to 0s and time points where the Anscombe residuals of the initial model are < 1.}
}}

