% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-DODmodel-class.R
\name{DODmodel}
\alias{DODmodel}
\title{Create a model for outbreak detection}
\usage{
DODmodel(family, formula, transMat = NULL, initProb = NULL, transMat_prior = TRUE, setBckgState = TRUE)
}
\arguments{
\item{family}{A DODfamily object specifying the family of probability distributions.}

\item{formula}{A DODformula object specifying the formula of the model.}

\item{transMat}{Inital transition probabilities.}

\item{transMat_prior}{Indicates whether a prior for transition probabilities should be used.}

\item{setBckgState}{Indicates whether a background state should be computed for model fitting. Background states are set to 0s and time points where the Anscombe residuals of the initial model are < 1.}
}
\description{
Create a model for outbreak detection
}
\examples{

TODO
  
}
\seealso{
\code{\linkS4class{DODmodel}}, \code{\linkS4class{DODfamily}}, \code{\linkS4class{DODformula}}
}
