% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Emission-class.R
\docType{class}
\name{Emission-class}
\alias{Emission-class}
\title{This class is a generic container for the model used for outbreak detection}
\description{
This class is a generic container for the model used for outbreak detection
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the emission.}
}}

