% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-DODformula-class.R
\docType{class}
\name{ExtData-class}
\alias{ExtData-class}
\title{This class is a container for the parameterization using external data.}
\description{
This class is a container for the parameterization using external data.
}
\section{Slots}{

\describe{
\item{\code{extdata}}{A data.frame containing variables which are used to model case counts. All variables in the data.frame will be used in the model. The data.frame has to have the same number of rows as time points in the time series.}

\item{\code{freq}}{Number of time units (can be e.g. weeks or months) in a year.}

\item{\code{offset}}{An offset to included in the predictor during fitting.}

\item{\code{formula_bckg}}{A formula which models the endemic (background) level of cases.}

\item{\code{formula}}{A formula which models the endemic (background) and epidemic levels of cases.}
}}

