% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoderNegBinom.R
\name{extractFeatures}
\alias{extractFeatures}
\title{Extract hidden representation of cases using a linear count Autoencoder}
\usage{
extractFeatures(K, offset_mat=matrix(1,nrow=nrow(K),ncol=ncol(K)), myrange=NULL, time_back=260, q=1, pseudocount=1, maxIter=1000, verbose=F)
}
\arguments{
\item{K}{Input matrix of case counts. Columns indicate regions (e.g. counties), rows indicate time points.}

\item{myrange}{Range of time points for which hidden representation should be calculated.}

\item{time_back}{Number of previous time points which should be included in the model.}

\item{q}{Number of hidden units. Default is 1.}

\item{pseudocount}{Number of pseudocounts to be added to case counts.}

\item{offset_mat}{An offset to included in the model.}

\item{maxIter}{Maximal number of iterations for model fitting.}

\item{verbose}{Indicates whether progress should be printed during model fitting.}
}
\description{
Extract hidden representation of cases using a linear count Autoencoder
}
\examples{

TODO
  
}
