% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoderNegBinom.R
\name{fitAutoencoder}
\alias{fitAutoencoder}
\title{Fit a linear count Autoencoder}
\usage{
fitAutoencoder(K, offset_mat=matrix(1,nrow=nrow(K),ncol=ncol(K)), q=1, pseudocount=1, maxIter=1000, verbose=F)
}
\arguments{
\item{K}{Input matrix of case counts. Columns indicate regions (e.g. counties), rows indicate time points.}

\item{offset_mat}{An offset to included in the model.}

\item{q}{Number of hidden units. Default is 1.}

\item{pseudocount}{Number of pseudocounts to be added to case counts.}

\item{maxIter}{Maximal number of iterations for model fitting.}

\item{verbose}{Indicates whether progress should be printed during model fitting.}
}
\description{
Fit a linear count Autoencoder
}
\examples{

TODO
  
}
