% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-DODformula-class.R
\docType{class}
\name{Harmonic-class}
\alias{Harmonic-class}
\title{This class is a container for the parameterization of the Harmonic models.}
\description{
This class is a container for the parameterization of the Harmonic models.
}
\section{Slots}{

\describe{
\item{\code{S}}{Number of oscillations during one year.}

\item{\code{timeTrend}}{Indicates whether a time trend should be included in the model.}

\item{\code{freq}}{Number of time units (can be e.g. weeks or months) in a year.}

\item{\code{offset}}{An offset to included in the predictor during fitting.}

\item{\code{formula_bckg}}{A formula which models the endemic (background) level of cases.}

\item{\code{formula}}{A formula which models the endemic (background) and epidemic levels of cases.}
}}

