% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot sts object with dod results}
\usage{
plot_model(dod_model, result, surv_ts, past_weeks_not_incuded=26, years_back=5)
}
\arguments{
\item{dod_model}{An object of class \code{\linkS4class{DODmodel}} which specifies the model parameters and structure}

\item{surv_ts}{A surveillance time series (sts) object created using the surveillance packages.}

\item{past_weeks_not_incuded}{Number of past weeks not included for initial glm esteimate.}

\item{years_back}{Number of past years to include in model fit.}

\item{results}{Result data frame of the dod function.}
}
\description{
Plot sts object with dod results
}
\examples{

TODO
  
}
