% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSts.R
\name{plotSts}
\alias{plotSts}
\title{Plot sts object with dod results}
\usage{
plotSts(surv_ts, results=NULL, range=NULL, main="", set_mar=TRUE, legend.offset=0, legend.offset_y=0, ...)
}
\arguments{
\item{results}{Result data frame of the dod function.}

\item{range}{Time points that should be plotted.}

\item{main}{Title of the plot.}

\item{set_mar}{Logical indicating whether to use margins mar=c(4,4,4,4) in plot.}

\item{legend.offset}{Position of Posterior legend along x-axis.}

\item{legend.offset_y}{Position of Posterior legend along y-axis.}

\item{surv_ts}{A surveillance time series (sts) object created using the surveillance packages.}
}
\description{
Plot sts object with dod results
}
\examples{

TODO
  
}
