% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-DODfamily-class.R
\docType{class}
\name{Poisson-class}
\alias{Poisson-class}
\title{This class is a generic container for a family of Poisson distributions}
\description{
This class is a generic container for a family of Poisson distributions
}
