% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-DODfamily-class.R
\docType{class}
\name{ZINegBinom-class}
\alias{ZINegBinom-class}
\title{This class is a generic container for a family of Negative Binomial distributions}
\description{
This class is a generic container for a family of Negative Binomial distributions
}
\section{Slots}{

\describe{
\item{\code{shared_dispersion}}{Logical indicating whether dispersion parameter is shared across multiple time series.}

\item{\code{shared_pi}}{Logical indicating whether parameter pi is shared across multiple time series.}
}}

