% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-DODfamily-class.R
\docType{class}
\name{ZIPoisson-class}
\alias{ZIPoisson-class}
\title{This class is a generic container for a family of Zero-Inflated Poisson distributions}
\description{
This class is a generic container for a family of Zero-Inflated Poisson distributions
}
\section{Slots}{

\describe{
\item{\code{shared_pi}}{Logical indicating whether parameter pi is shared across multiple time series.}
}}

